pyproject.toml
[project]
name = ""
version = ""
description = "GPTStonks API allows interacting with financial data sources using natural language."
authors = [
    {name = "GPTStonks", email = "gptstonks@gmail.com"},
]
dependencies = [
    "pre-commit>=3.6.0",
    "fastapi>=0.104.1",
    "openbb==4.1.2",
    "openbb-chat>=0.0.9",
    "uvicorn>=0.23.2",
    "sentence-transformers>=2.2.2",
    "duckduckgo-search>=3.8.1",
    "wikipedia>=1.4.0",
    "youtube-search>=2.1.2",
    "yfinance>=0.2.31",
    "gdown>=4.7.1",
    "boto3>=1.33.6",
    "langchain>=0.0.342",
    "llama-index>=0.9.13",
    "python-dotenv>=1.0.0",
    "openbb-alpha-vantage>=1.1.0",
    "openbb-biztoc>=1.1.0",
    "openbb-cboe>=1.1.0",
    "openbb-ecb>=1.1.0",
    "openbb-finra>=1.1.0",
    "openbb-government-us>=1.1.0",
    "openbb-nasdaq>=1.1.1",
    "openbb-seeking-alpha>=1.1.0",
    "openbb-stockgrid>=1.1.0",
    "openbb-wsj>=1.1.0",
    "pymongo>=4.6.1",
    "langchain-openai>=0.0.2.post1",
]
requires-python = ">=3.10,<3.11"
readme = "README.md"
license = {text = "MIT"}

[project.optional-dependencies]
testing = [
    "pytest",
    "pytest-cov",
    "trio>=0.23.2",
    "pytest-asyncio>=0.23.2",
]
llamacpp = [
    "llama-cpp-python>=0.2.23",
]
docs = [
    "mkdocs>=1.5.3",
    "mkdocstrings[python]>=0.24.0",
    "mkdocs-material[imaging]>=9.5.9",
    "mkdocs-gen-files>=0.5.0",
    "mkdocs-literate-nav>=0.6.1",
    "mkdocs-minify-plugin>=0.8.0",
    "mkdocs-git-revision-date-localized-plugin>=1.2.4",
    "mkdocs-section-index>=0.3.8",
    "essentials-openapi[full]>=1.0.9",
]

# [tool.pdm.resolution.overrides]
# Not needed for now

[tool.pytest.ini_options]
addopts = [
    "--import-mode=importlib",
]

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"
